Quote fetch service (unstable).

+++ src/main/groovy/com/gastocks/server/models/Quote.groovy
@@ -14,7 +14,7 @@ class Quote {
     Double currentTradingDayLow
     Double previousTradingDayClose
     Double priceChange
-    Double priceChangePercentage
+    Float priceChangePercentage
     Integer volume
     DateTime lastUpdated
